ER-629 Renamed and moved VacancyReviewSearch (now QaVacancySummary) from domain to infrastructure projections.
