+++ src/Shared/Recruit.Vacancies.Client/Infrastructure/Client/IQaVacancyClient.cs
@@ -18,6 +18,6 @@ namespace Esfa.Recruit.Vacancies.Client.Infrastructure.Client
         Task ReferVacancyReviewAsync(Guid reviewId);
         Task ApproveReferredReviewAsync(Guid reviewId, string shortDescription, string vacancyDescription, string trainingDescription, string outcomeDescription, string thingsToConsider, string employerDescription);
         Task<Qualifications> GetCandidateQualificationsAsync();
-        Task<List<VacancyReviewSearch>> GetSearchResultsAsync(string searchTerm);
+        Task<List<QaVacancySummary>> GetSearchResultsAsync(string searchTerm);
     }
 }
