+++ src/Shared/Recruit.Vacancies.Client/Infrastructure/Client/QaVacancyClient.cs
@@ -83,13 +83,13 @@ namespace Esfa.Recruit.Vacancies.Client.Infrastructure.Client
             return dashboard;
         }
 
-        public async Task<List<VacancyReviewSearch>> GetSearchResultsAsync(string searchTerm)
+        public async Task<List<QaVacancySummary>> GetSearchResultsAsync(string searchTerm)
         {
             if (TryGetVacancyReference(searchTerm, out var vacancyReference))
             {
                 return await _reviewRepository.SearchAsync(vacancyReference);
             }
-            return new List<VacancyReviewSearch>();
+            return new List<QaVacancySummary>();
         }
 
         private static bool TryGetVacancyReference(string value, out long vacancyReference)
